First major refactor, WIP
