+++ reVX/utilities/__init__.py
@@ -2,7 +2,7 @@
 """
 reVX utilities.
 """
-from .exclusions_converter import ExclusionsConverter
+from .exclusions import ExclusionsConverter
 from .forecasts import FcstUtils
 from .output_extractor import output_extractor
 from .region_classifier import RegionClassifier
