+++ reVX/utilities/exclusions.py
@@ -1,0 +1,1457 @@
+# -*- coding: utf-8 -*-
+"""
+Driver class to compute exclusions
+"""
+import h5py
+import json
+import os
+import logging
+from abc import ABC, abstractmethod
+from concurrent.futures import as_completed
+from warnings import warn
+
+import numpy as np
+import geopandas as gpd
+from pyproj.crs import CRS
+import rasterio
+from rasterio import features
+from shapely.geometry import shape
+
+from rex import Outputs
+from rex.utilities import SpawnProcessPool, log_mem, parse_table
+from reV.handlers.exclusions import ExclusionLayers
+from reVX.handlers.geotiff import Geotiff
+from reVX.utilities.utilities import log_versions
+from reVX.utilities.exceptions import ExclusionsCheckError
+
+logger = logging.getLogger(__name__)
+
+
+class AbstractExclusionCalculatorInterface(ABC):
+    """Abstract Exclusion Calculator Interface. """
+
+    @property
+    @abstractmethod
+    def profile(self):
+        """dict: Geotiff profile. """
+        raise NotImplementedError
+
+    @abstractmethod
+    def parse_features(self, features_fpath):
+        """Parse features the feature file.
+
+        Parameters
+        ----------
+        features_fpath : str
+            Path to file containing features to compute exclusions from.
+
+        Returns
+        -------
+        `geopandas.GeoDataFrame`
+            Geometries of features to compute exclusions from in
+            exclusion coordinate system.
+        """
+        raise NotImplementedError
+
+    @abstractmethod
+    def pre_process_regulations(self, features_fpath):
+        """Reduce regulations to correct state and features.
+
+        When implementing this method, make sure to update
+        `self.regulations_table`.
+
+        Parameters
+        ----------
+        features_fpath : str
+            Path to shape file with features to compute exclusions from.
+        """
+        raise NotImplementedError
+
+    @abstractmethod
+    def compute_local_exclusions(self, features, cnty, regulation_value):
+        """Compute local feature exclusions.
+
+        This method should compute the exclusions using the information
+        about the input county.
+
+        Parameters
+        ----------
+        features : geopandas.GeoDataFrame
+            Features used to calculate exclusions from.
+        cnty : geopandas.GeoDataFrame
+            Regulations for a single county.
+        regulation_value : int
+            Regulation value for county.
+
+        Returns
+        -------
+        exclusions : list
+            List of exclusion geometries.
+        """
+        raise NotImplementedError
+
+    @abstractmethod
+    def compute_generic_exclusions(self, features_fpath):
+        """Compute generic exclusions.
+
+        This method should compute the exclusions using a generic
+        regulation value (`self._regulations.generic`).
+
+        Parameters
+        ----------
+        features_fpath : str
+            Path to shape file with features to compute exclusions from.
+
+        Returns
+        -------
+        exclusions : ndarray
+            Raster array of exclusions
+        """
+        raise NotImplementedError
+
+    @staticmethod
+    @abstractmethod
+    def get_feature_paths(features_fpath):
+        """Ensure features path exists and return as list.
+
+        Parameters
+        ----------
+        features_fpath : str
+            Path to features file. This path can contain
+            any pattern that can be used in the glob function.
+            For example, `/path/to/features/[A]*` would match
+            with all the features in the directory
+            `/path/to/features/` that start with "A". This input
+            can also be a directory, but that directory must ONLY
+            contain feature files. If your feature files are mixed
+            with other files or directories, use something like
+            `/path/to/features/*.geojson`.
+
+        Returns
+        -------
+        features_fpath : list
+            Features path as a list of strings.
+
+        Notes
+        -----
+        This method is required for `run` classmethods for exclusion
+        features that are spread out over multiple files.
+        """
+        raise NotImplementedError
+
+
+class AbstractBaseExclusionsMerger(AbstractExclusionCalculatorInterface):
+    """
+    Create exclusions layers for exclusions
+    """
+
+    def __init__(self, excl_fpath, regulations):
+        """
+        Parameters
+        ----------
+        excl_fpath : str
+            Path to .h5 file containing exclusion layers, will also be
+            the location of any new exclusion layers
+        regulations : `Regulations`
+            A `Regulations` object used to extract exclusion regulation
+            values.
+        """
+        log_versions(logger)
+        self._excl_fpath = excl_fpath
+        self._regulations = regulations
+        with ExclusionLayers(self._excl_fpath) as exc:
+            self._fips = exc['cnty_fips']
+            self._cnty_fips_profile = exc.get_layer_profile('cnty_fips')
+        self._preflight_check()
+
+    def __repr__(self):
+        msg = "{} for {}".format(self.__class__.__name__, self._excl_fpath)
+        return msg
+
+    def _preflight_check(self):
+        """Parse the county regulations.
+
+        Parse regulations, combine with county geometries from
+        exclusions .h5 file. The county geometries are intersected with
+        features to compute county specific exclusions.
+
+        Parameters
+        ----------
+        regulations : pandas.DataFrame
+            Regulations table
+
+        Returns
+        -------
+        regulations: `geopandas.GeoDataFrame`
+            GeoDataFrame with county level exclusion regulations merged
+            with county geometries, use for intersecting with exclusion
+            features.
+        """
+        if self.regulations_table is None:
+            return
+
+        regulations_df = self.regulations_table
+        if 'FIPS' not in regulations_df:
+            msg = ('Regulations does not have county FIPS! Please add a '
+                   '"FIPS" columns with the unique county FIPS values.')
+            logger.error(msg)
+            raise RuntimeError(msg)
+
+        if 'geometry' not in regulations_df:
+            regulations_df['geometry'] = None
+
+        regulations_df = regulations_df[~regulations_df['FIPS'].isna()]
+        regulations_df = regulations_df.set_index('FIPS')
+
+        logger.info('Merging county geometries w/ local regulations')
+        s = features.shapes(
+            self._fips.astype(np.int32),
+            transform=self._cnty_fips_profile['transform']
+        )
+        for p, v in s:
+            v = int(v)
+            if v in regulations_df.index:
+                regulations_df.at[v, 'geometry'] = shape(p)
+
+        regulations_df = gpd.GeoDataFrame(
+            regulations_df,
+            crs=self.profile['crs'],
+            geometry='geometry'
+        )
+        regulations_df = regulations_df.reset_index()
+        regulations_df = regulations_df.to_crs(crs=self.profile['crs'])
+        self.regulations_table = regulations_df
+
+    @property
+    def regulations_table(self):
+        """Regulations table.
+
+        Returns
+        -------
+        geopandas.GeoDataFrame | None
+        """
+        return self._regulations.regulations
+
+    @regulations_table.setter
+    def regulations_table(self, regulations_table):
+        self._regulations.regulations = regulations_table
+
+    # def _parse_features(self, features_fpath):
+    #     """Method to parse features.
+
+    #     Parameters
+    #     ----------
+    #     features_fpath : str
+    #         Path to file containing features to setback from.
+
+    #     Returns
+    #     -------
+    #     `geopandas.GeoDataFrame`
+    #         Geometries of features to setback from in exclusion
+    #         coordinate system.
+    #     """
+    #     return gpd.read_file(features_fpath).to_crs(
+    #         crs=self._rasterizer.profile['crs'])
+
+    # def _pre_process_regulations(self, features_fpath):
+    #     """Reduce regulations to state corresponding to features_fpath.
+
+    #     Parameters
+    #     ----------
+    #     features_fpath : str
+    #         Path to shape file with features to compute setbacks from.
+    #     """
+    #     mask = self._regulation_table_mask(features_fpath)
+    #     if not mask.any():
+    #         msg = "Found no local regulations!"
+    #         logger.warning(msg)
+    #         warn(msg)
+
+    #     self.regulations_table = (self.regulations_table[mask]
+    #                               .reset_index(drop=True))
+    #     logger.debug('Computing setbacks for regulations in {} counties'
+    #                  .format(len(self.regulations_table)))
+
+    # pylint: disable=unused-argument
+    # @abstractmethod
+    # def _regulation_table_mask(self, features_fpath):
+    #     """Return the regulation table mask for setback feature. """
+    #     raise NotImplementedError
+
+    # def _compute_local_setbacks(self, features, cnty, setback):
+    #     """Compute local features setbacks.
+
+    #     This method will compute the setbacks using a county-specific
+    #     regulations file that specifies either a static setback or a
+    #     multiplier value that will be used along with the base setback
+    #     distance to compute the setback.
+
+    #     Parameters
+    #     ----------
+    #     features : geopandas.GeoDataFrame
+    #         Features to setback from.
+    #     cnty : geopandas.GeoDataFrame
+    #         Regulations for a single county.
+    #     setback : int
+    #         Setback distance in meters.
+
+    #     Returns
+    #     -------
+    #     setbacks : list
+    #         List of setback geometries.
+    #     """
+        # logger.debug('- Computing setbacks for county FIPS {}'
+        #              .format(cnty.iloc[0]['FIPS']))
+        # log_mem(logger)
+        # features = self._feature_filter(features, cnty)
+        # return list(features.buffer(setback))
+        # TODO: Delegate this to class
+
+    # @staticmethod
+    # def _feature_filter(features, cnty):
+    #     """Filter the features given a county."""
+    #     return features_with_centroid_in_county(features, cnty)
+
+    def _write_exclusions(self, geotiff, exclusions, replace=False):
+        """
+        Write exclusions to geotiff, replace if requested
+
+        Parameters
+        ----------
+        geotiff : str
+            Path to geotiff file to save exclusions too
+        exclusions : ndarray
+            Rasterized array of exclusions.
+        replace : bool, optional
+            Flag to replace local layer data with arr if layer already
+            exists in the exclusion .h5 file. By default `False`.
+        """
+        if os.path.exists(geotiff):
+            if not replace:
+                msg = ('{} already exists. To replace it set "replace=True"'
+                       .format(geotiff))
+                logger.error(msg)
+                raise IOError(msg)
+            else:
+                msg = ('{} already exists and will be replaced!'
+                       .format(geotiff))
+                logger.warning(msg)
+                warn(msg)
+
+        ExclusionsConverter._write_geotiff(geotiff, self.profile, exclusions)
+
+    def compute_all_local_exclusions(self, features_fpath, max_workers=None):
+        """Compute local exclusions for all counties either.
+
+        Parameters
+        ----------
+        features_fpath : str
+            Path to shape file with features to compute exclusions from
+        max_workers : int, optional
+            Number of workers to use for exclusions computation, if 1
+            run in serial, if > 1 run in parallel with that many
+            workers, if `None` run in parallel on all available cores.
+            By default `None`.
+
+        Returns
+        -------
+        exclusions : ndarray
+            Raster array of exclusions.
+        """
+        exclusions = None
+        features = self.parse_features(features_fpath)
+        max_workers = max_workers or os.cpu_count()
+
+        log_mem(logger)
+        if max_workers > 1:
+            logger.info('Computing local exclusions in parallel using {} '
+                        'workers'.format(max_workers))
+            loggers = [__name__, 'reVX']
+            with SpawnProcessPool(max_workers=max_workers,
+                                  loggers=loggers) as exe:
+                futures = {}
+                for func, *args in self._exclusions_computation(features):
+                    cnty_feats, cnty, exclusion = args
+                    future = exe.submit(func, cnty_feats, cnty, exclusion)
+                    futures[future] = cnty['FIPS'].unique()
+
+                for i, future in enumerate(as_completed(futures)):
+                    exclusions = self._combine_exclusions(exclusions,
+                                                          future.result(),
+                                                          futures[future])
+                    logger.debug('Computed exclusions for {} of {} counties'
+                                 .format((i + 1), len(self.regulations_table)))
+        else:
+            logger.info('Computing local exclusions in serial')
+            computation = self._exclusions_computation(features)
+            for i, (func, *args) in enumerate(computation):
+                cnty_feats, cnty, exclusion = args
+                exclusions = self._combine_exclusions(exclusions, func(*args),
+                                                      cnty['FIPS'].unique())
+                logger.debug('Computed exclusions for {} of {} counties'
+                             .format((i + 1), len(self.regulations_table)))
+
+        return exclusions
+
+    def _exclusions_computation(self, features):
+        """Get function and args for exclusions computation. """
+        for exclusion, cnty in self._regulations:
+            idx = features.sindex.intersection(cnty.total_bounds)
+            cnty_feats = features.iloc[list(idx)].copy()
+            yield self.compute_local_exclusions, cnty_feats, cnty, exclusion
+
+    # def _compute_generic_setbacks(self, features_fpath):
+    #     """Compute generic setbacks.
+
+    #     This method will compute the setbacks using a generic setback
+    #     of `base_setback_dist * multiplier`.
+
+    #     Parameters
+    #     ----------
+    #     features_fpath : str
+    #         Path to shape file with features to compute setbacks from.
+
+    #     Returns
+    #     -------
+    #     setbacks : ndarray
+    #         Raster array of setbacks
+    #     """
+        # logger.info('Computing generic setbacks')
+        # if np.isclose(self._regulations.generic_setback, 0):
+        #     return self._rasterizer.rasterize(shapes=None)
+
+        # setback_features = self._parse_features(features_fpath)
+        # setbacks = list(setback_features.buffer(
+        #     self._regulations.generic_setback
+        # ))
+
+        # return self._rasterizer.rasterize(setbacks)
+        # TODO: Delegate this to class
+
+    def compute_exclusions(self, features_fpath, max_workers=None,
+                           geotiff=None, replace=False):
+        """
+        Compute exclusions for all states either in serial or parallel.
+        Existing exclusions are computed if a regulations file was
+        supplied during class initialization, otherwise generic exclusions
+        are computed.
+
+        Parameters
+        ----------
+        features_fpath : str
+            Path to shape file with features to compute exclusions from
+        max_workers : int, optional
+            Number of workers to use for exclusion computation, if 1 run
+            in serial, if > 1 run in parallel with that many workers,
+            if `None`, run in parallel on all available cores.
+            By default `None`.
+        geotiff : str, optional
+            Path to save geotiff containing rasterized exclusions.
+            By default `None`.
+        replace : bool, optional
+            Flag to replace geotiff if it already exists.
+            By default `False`.
+
+        Returns
+        -------
+        exclusions : ndarray
+            Raster array of exclusions
+        """
+        exclusions = self._compute_merged_exclusions(features_fpath,
+                                                     max_workers=max_workers)
+
+        if geotiff is not None:
+            logger.debug('Writing exclusions to {}'.format(geotiff))
+            self._write_exclusions(geotiff, exclusions, replace=replace)
+
+        return exclusions
+
+    def _compute_merged_exclusions(self, features_fpath, max_workers=None):
+        """Compute and merge local and generic exclusions, if necessary. """
+        mw = max_workers
+
+        if self._regulations.locals_exist:
+            self.pre_process_regulations(features_fpath)
+
+        generic_exclusions_exist = self._regulations.generic_exists
+        local_exclusions_exist = self._regulations.locals_exist
+
+        if not generic_exclusions_exist and not local_exclusions_exist:
+            msg = ("Found no exclusions to compute: No regulations detected, "
+                   "and generic multiplier not set.")
+            logger.error(msg)
+            raise ValueError(msg)
+
+        if generic_exclusions_exist and not local_exclusions_exist:
+            return self.compute_generic_exclusions(features_fpath)
+
+        if local_exclusions_exist and not generic_exclusions_exist:
+            return self.compute_all_local_exclusions(features_fpath,
+                                                     max_workers=mw)
+
+        generic_exclusions = self.compute_generic_exclusions(features_fpath)
+        local_exclusions = self.compute_all_local_exclusions(features_fpath,
+                                                             max_workers=mw)
+        return self._merge_exclusions(generic_exclusions, local_exclusions,
+                                      features_fpath)
+
+    def _merge_exclusions(self, generic_exclusions, local_exclusions,
+                          features_fpath):
+        """Merge local exclusions onto the generic exclusions."""
+        logger.info('Merging local exclusions onto the generic exclusions')
+
+        self.pre_process_regulations(features_fpath)
+        local_fips = self.regulations_table["FIPS"].unique()
+        return self._combine_exclusions(generic_exclusions, local_exclusions,
+                                        local_fips)
+
+    def _combine_exclusions(self, existing, additional, cnty_fips):
+        """Combine local exclusions using FIPS code"""
+        if existing is None:
+            return additional
+
+        local_exclusions_mask = np.isin(self._fips, cnty_fips)
+        existing[local_exclusions_mask] = additional[local_exclusions_mask]
+        return existing
+
+    @classmethod
+    def run(cls, excl_fpath, features_path, out_dir, regulations,
+            weights_calculation_upscale_factor=None, max_workers=None,
+            replace=False, hsds=False):
+        """
+        Compute exclusions and write them to a geotiff. If a regulations
+        file is given, compute local exclusions, otherwise compute
+        generic exclusions. If both are provided, generic and local
+        exclusions are merged such that the local exclusions override
+        the generic ones.
+
+        Parameters
+        ----------
+        excl_fpath : str
+            Path to .h5 file containing exclusion layers, will also be
+            the location of any new exclusion layers.
+        features_path : str
+            Path to file or directory feature shape files.
+            This path can contain any pattern that can be used in the
+            glob function. For example, `/path/to/features/[A]*` would
+            match with all the features in the directory
+            `/path/to/features/` that start with "A". This input
+            can also be a directory, but that directory must ONLY
+            contain feature files. If your feature files are mixed
+            with other files or directories, use something like
+            `/path/to/features/*.geojson`.
+        out_dir : str
+            Directory to save exclusion geotiff(s) into
+        regulations : `Regulations`
+            A `Regulations` object used to extract exclusion regulation
+            distances.
+        weights_calculation_upscale_factor : int, optional
+            If this value is an int > 1, the output will be a layer with
+            **inclusion** weight values instead of exclusion booleans.
+            For example, a cell that was previously excluded with a
+            a boolean mask (value of 1) may instead be converted to an
+            inclusion weight value of 0.75, meaning that 75% of the area
+            corresponding to that point should be included (i.e. the
+            exclusion feature only intersected a small portion - 25% -
+            of the cell). This percentage inclusion value is calculated
+            by upscaling the output array using this input value,
+            rasterizing the exclusion features onto it, and counting the
+            number of resulting sub-cells excluded by the feature. For
+            example, setting the value to `3` would split each output
+            cell into nine sub-cells - 3 divisions in each dimension.
+            After the feature is rasterized on this high-resolution
+            sub-grid, the area of the non-excluded sub-cells is totaled
+            and divided by the area of the original cell to obtain the
+            final inclusion percentage. Therefore, a larger upscale
+            factor results in more accurate percentage values. However,
+            this process is memory intensive and scales quadratically
+            with the upscale factor. A good way to estimate your minimum
+            memory requirement is to use the following formula:
+
+            .. math:: memory (GB) = s_0 * s_1 * ((sf^2) * 2 + 4) / 1073741824,
+
+            where :math:`s_0` and :math:`s_1` are the dimensions (shape)
+            of your exclusion layer and :math:`sf` is the scale factor
+            (be sure to add several GB for any other overhead required
+            by the rest of the process). If `None` (or a value <= 1),
+            this process is skipped and the output is a boolean
+            exclusion mask. By default `None`.
+        max_workers : int, optional
+            Number of workers to use for exclusion computation, if 1 run
+            in serial, if > 1 run in parallel with that many workers,
+            if `None`, run in parallel on all available cores.
+            By default `None`.
+        replace : bool, optional
+            Flag to replace geotiff if it already exists.
+            By default `False`.
+        hsds : bool, optional
+            Boolean flag to use h5pyd to handle .h5 'files' hosted on
+            AWS behind HSDS. By default `False`.
+        """
+        scale_factor = weights_calculation_upscale_factor
+        exclusions = cls(excl_fpath, regulations=regulations, hsds=hsds,
+                         weights_calculation_upscale_factor=scale_factor)
+
+        features_path = exclusions.get_feature_paths(features_path)
+        for fpath in features_path:
+            geotiff = os.path.basename(fpath)
+            geotiff = ".".join(geotiff.split('.')[:-1] + ['tif'])
+            geotiff = os.path.join(out_dir, geotiff)
+
+            if os.path.exists(geotiff) and not replace:
+                msg = ('{} already exists, exclusions will not be re-computed '
+                       'unless replace=True'.format(geotiff))
+                logger.error(msg)
+            else:
+                logger.info("Computing exclusions from {} and saving "
+                            "to {}".format(fpath, geotiff))
+                exclusions.compute_exclusions(fpath, geotiff=geotiff,
+                                              max_workers=max_workers,
+                                              replace=replace)
+
+
+class ExclusionsConverter:
+    """
+    Convert exclusion layers between .h5 and .tif (geotiff)
+    """
+    def __init__(self, excl_h5, hsds=False, chunks=(128, 128), replace=True):
+        """
+        Parameters
+        ----------
+        excl_h5 : str
+            Path to .h5 file containing or to contain exclusion layers
+        hsds : bool, optional
+            Boolean flag to use h5pyd to handle .h5 'files' hosted on AWS
+            behind HSDS, by default False
+        chunks : tuple, optional
+            Chunk size of exclusions in .h5 and Geotiffs, by default (128, 128)
+        replace : bool, optional
+            Flag to replace existing layers if needed, by default True
+        """
+        log_versions(logger)
+        self._excl_h5 = excl_h5
+        self._hsds = hsds
+        self._chunks = chunks
+        self._replace = replace
+
+    def __repr__(self):
+        msg = "{} for {}".format(self.__class__.__name__, self._excl_h5)
+        return msg
+
+    def __getitem__(self, layer):
+        """
+        Parameters
+        ----------
+        layer : str
+            Layer to extract data for
+
+        Returns
+        -------
+        profile : dict
+            Geotiff profile (attributes)
+        values : ndarray
+            Geotiff data
+        """
+
+        if layer not in self.layers:
+            msg = "{} is not present in {}".format(layer, self._excl_h5)
+            logger.error(msg)
+            raise KeyError(msg)
+
+        profile, values = self._extract_layer(self._excl_h5, layer,
+                                              hsds=self._hsds)
+        return profile, values
+
+    def __setitem__(self, layer, geotiff):
+        """
+        Parameters
+        ----------
+        layer : str
+            Layer to set
+        geotiff : str
+            Path to GeoTiff to load data from
+        """
+        self.geotiff_to_layer(layer, geotiff)
+
+    @property
+    def layers(self):
+        """
+        Available exclusion layers in .h5 file
+
+        Returns
+        -------
+        layers : list
+            Available layers in .h5 file
+        """
+        with ExclusionLayers(self._excl_h5, hsds=self._hsds) as exc:
+            layers = exc.layers
+
+        return layers
+
+    @staticmethod
+    def _init_h5(excl_h5, geotiff, chunks=(128, 128)):
+        """
+        Initialize exclusions .h5 file from geotiff:
+        - Transfer profile, shape, and meta
+
+        Parameters
+        ----------
+        excl_h5 : str
+            Path to .h5 file containing exclusion layers
+        geotiff : str
+            Path to geotiff file
+        chunks : tuple
+            Chunk size of exclusions in Geotiff
+        """
+        logger.debug('\t- Initializing {} from {}'
+                     .format(excl_h5, geotiff))
+        with Geotiff(geotiff, chunks=chunks) as src:
+            profile = src.profile
+            shape = src.shape
+            lat, lon = src.lat_lon
+            logger.debug('\t- "profile", "meta", and "shape" extracted from {}'
+                         .format(geotiff))
+
+        try:
+            with h5py.File(excl_h5, mode='w') as dst:
+                dst.attrs['profile'] = json.dumps(profile)
+                logger.debug('\t- Default profile:\n{}'.format(profile))
+                dst.attrs['shape'] = shape
+                logger.debug('\t- Default shape:\n{}'.format(shape))
+                dst.attrs['chunks'] = chunks
+                logger.debug('\t- Default chunks:\n{}'.format(chunks))
+
+                dst.create_dataset('latitude', shape=lat.shape,
+                                   dtype=np.float32, data=lat,
+                                   chunks=chunks)
+                logger.debug('\t- latitude coordiantes created')
+
+                dst.create_dataset('longitude', shape=lon.shape,
+                                   dtype=np.float32, data=lon,
+                                   chunks=chunks)
+                logger.debug('\t- longitude coordiantes created')
+        except Exception:
+            logger.exception("Error initilizing {}".format(excl_h5))
+            if os.path.exists(excl_h5):
+                os.remove(excl_h5)
+
+    @staticmethod
+    def _check_crs(baseline_crs, test_crs, ignore_keys=('no_defs',)):
+        """
+        Compare baseline and test crs values
+
+        Parameters
+        ----------
+        baseline_crs : dict
+            Baseline CRS to use a truth, must be a dict
+        test_crs : dict
+            Test CRS to compare with baseline, must be a dictionary
+        ignore_keys : tuple
+            Keys to not check
+
+        Returns
+        -------
+        bad_crs : bool
+            Flag if crs' do not match
+        """
+        bad_crs = False
+        for k, true_v in baseline_crs.items():
+            if k not in ignore_keys:
+                test_v = test_crs.get(k, true_v)
+                if true_v != test_v:
+                    bad_crs = True
+
+        return bad_crs
+
+    @classmethod
+    def _check_geotiff(cls, excl_h5, geotiff, chunks=(128, 128),
+                       transform_atol=0.01, coord_atol=0.001):
+        """
+        Compare geotiff with exclusion layer, raise any errors
+
+        Parameters
+        ----------
+        excl_h5 : str
+            Path to .h5 file containing exclusion layers
+        geotiff : str
+            Path to geotiff file
+        chunks : tuple
+            Chunk size of exclusions in Geotiff
+        transform_atol : float
+            Absolute tolerance parameter when comparing geotiff transform data.
+        coord_atol : float
+            Absolute tolerance parameter when comparing new un-projected
+            geotiff coordinates against previous coordinates.
+        """
+        with Geotiff(geotiff, chunks=chunks) as tif:
+            with ExclusionLayers(excl_h5) as h5:
+                if tif.bands > 1:
+                    error = ('{} contains more than one band!'
+                             .format(geotiff))
+                    logger.error(error)
+                    raise ExclusionsCheckError(error)
+
+                if not np.array_equal(h5.shape, tif.shape):
+                    error = ('Shape of exclusion data in {} and {} do not '
+                             'match!'.format(geotiff, excl_h5))
+                    logger.error(error)
+                    raise ExclusionsCheckError(error)
+
+                profile = h5.profile
+                h5_crs = CRS.from_string(profile['crs']).to_dict()
+                tif_crs = CRS.from_string(tif.profile['crs']).to_dict()
+                bad_crs = cls._check_crs(h5_crs, tif_crs)
+                if bad_crs:
+                    error = ('Geospatial "crs" in {} and {} do not match!'
+                             '\n {} !=\n {}'
+                             .format(geotiff, excl_h5, tif_crs, h5_crs))
+                    logger.error(error)
+                    raise ExclusionsCheckError(error)
+
+                if not np.allclose(profile['transform'],
+                                   tif.profile['transform'],
+                                   atol=transform_atol):
+                    error = ('Geospatial "transform" in {} and {} do not '
+                             'match!\n {} !=\n {}'
+                             .format(geotiff, excl_h5, profile['transform'],
+                                     tif.profile['transform']))
+                    logger.error(error)
+                    raise ExclusionsCheckError(error)
+
+                lat, lon = tif.lat_lon
+                if not np.allclose(h5.latitude, lat, atol=coord_atol):
+                    error = ('Latitude coordinates {} and {} do not match to '
+                             'within {} degrees!'
+                             .format(geotiff, excl_h5, coord_atol))
+                    logger.error(error)
+                    raise ExclusionsCheckError(error)
+
+                if not np.allclose(h5.longitude, lon, atol=coord_atol):
+                    error = ('Longitude coordinates {} and {} do not match to '
+                             'within {} degrees!'
+                             .format(geotiff, excl_h5, coord_atol))
+                    logger.error(error)
+                    raise ExclusionsCheckError(error)
+
+    @classmethod
+    def _parse_tiff(cls, geotiff, excl_h5=None, chunks=(128, 128),
+                    check_tiff=True, transform_atol=0.01, coord_atol=0.001):
+        """
+        Extract exclusion layer from given geotiff, compare with excl_h5
+        if provided
+
+        Parameters
+        ----------
+        geotiff : str
+            Path to geotiff file
+        excl_h5 : str, optional
+            Path to .h5 file containing exclusion layers, by default None
+        chunks : tuple, optional
+            Chunk size of exclusions in Geotiff, by default (128, 128)
+        check_tiff : bool, optional
+            Flag to check tiff profile and coordinates against exclusion .h5
+            profile and coordinates, by default True
+        transform_atol : float, optional
+            Absolute tolerance parameter when comparing geotiff transform data,
+            by default 0.01
+        coord_atol : float, optional
+            Absolute tolerance parameter when comparing new un-projected
+            geotiff coordinates against previous coordinates, by default 0.001
+
+        Returns
+        -------
+        profile : dict
+            Geotiff profile (attributes)
+        values : ndarray
+            Geotiff data
+        """
+        if excl_h5 is not None and check_tiff:
+            cls._check_geotiff(excl_h5, geotiff, chunks=chunks,
+                               transform_atol=transform_atol,
+                               coord_atol=coord_atol)
+
+        with Geotiff(geotiff, chunks=chunks) as tif:
+            profile, values = tif.profile, tif.values
+
+        return profile, values
+
+    @staticmethod
+    def _write_layer(excl_h5, layer, profile, values, chunks=(128, 128),
+                     description=None, scale_factor=None):
+        """
+        Write exclusion layer to .h5 file
+
+        Parameters
+        ----------
+        excl_h5 : str
+            Path to .h5 file containing exclusion layers
+        layer : str
+            Dataset name in .h5 file
+        profile : dict
+            Geotiff profile (attributes)
+        values : ndarray
+            Geotiff data
+        chunks : tuple
+            Chunk size of dataset in .h5 file
+        description : str
+            Description of exclusion layer
+        scale_factor : int | float, optional
+            Scale factor to use to scale geotiff data when added to the .h5
+            file, by default None
+        """
+        if len(chunks) < 3:
+            chunks = (1, ) + chunks
+
+        if values.ndim < 3:
+            values = np.expand_dims(values, 0)
+
+        with h5py.File(excl_h5, mode='a') as f:
+            if layer in f:
+                ds = f[layer]
+                ds[...] = values
+                logger.debug('\t- {} values replaced'.format(layer))
+            else:
+                ds = f.create_dataset(layer, shape=values.shape,
+                                      dtype=values.dtype, chunks=chunks,
+                                      data=values)
+                logger.debug('\t- {} created and loaded'.format(layer))
+
+            ds.attrs['profile'] = json.dumps(profile)
+            logger.debug('\t- Unique profile for {} added:\n{}'
+                         .format(layer, profile))
+            if description is not None:
+                ds.attrs['description'] = description
+                logger.debug('\t- Description for {} added:\n{}'
+                             .format(layer, description))
+
+            if scale_factor is not None:
+                ds.attrs['scale_factor'] = scale_factor
+                logger.debug('\t- scale_factor for {} added:\n{}'
+                             .format(layer, scale_factor))
+
+    @classmethod
+    def _geotiff_to_h5(cls, excl_h5, layer, geotiff, chunks=(128, 128),
+                       check_tiff=True, transform_atol=0.01, coord_atol=0.001,
+                       description=None, scale_factor=None, dtype='int16'):
+        """
+        Transfer geotiff exclusions to h5 confirming they match existing layers
+
+        Parameters
+        ----------
+        excl_h5 : str
+            Path to .h5 file containing exclusion layers
+        layer : str
+            Layer to extract
+        geotiff : str
+            Path to geotiff file
+        chunks : tuple, optional
+            Chunk size of exclusions in Geotiff, by default (128, 128)
+        check_tiff : bool, optional
+            Flag to check tiff profile and coordinates against exclusion .h5
+            profile and coordinates, by default True
+        transform_atol : float, optional
+            Absolute tolerance parameter when comparing geotiff transform data,
+            by default 0.01
+        coord_atol : float, optional
+            Absolute tolerance parameter when comparing new un-projected
+            geotiff coordinates against previous coordinates, by default 0.001
+        description : str, optional
+            Description of exclusion layer, by default None
+        scale_factor : int | float, optional
+            Scale factor to use to scale geotiff data when added to the .h5
+            file, by default None
+        dtype : str, optional
+            Dtype to save geotiff data as in the .h5 file. Only used when
+            'scale_factor' is not None, by default 'int16'
+        """
+        logger.debug('\t- {} being extracted from {} and added to {}'
+                     .format(layer, geotiff, os.path.basename(excl_h5)))
+
+        profile, values = cls._parse_tiff(
+            geotiff, excl_h5=excl_h5, chunks=chunks, check_tiff=check_tiff,
+            transform_atol=transform_atol, coord_atol=coord_atol)
+
+        if scale_factor is not None:
+            attrs = {'scale_factor': scale_factor}
+            values = Outputs._check_data_dtype(layer, values, dtype,
+                                               attrs=attrs)
+
+        cls._write_layer(excl_h5, layer, profile, values,
+                         chunks=chunks, description=description,
+                         scale_factor=scale_factor)
+
+    @staticmethod
+    def _write_geotiff(geotiff, profile, values):
+        """
+        Write values to geotiff with given profile
+
+        Parameters
+        ----------
+        geotiff : str
+            Path to geotiff file to save data to
+        profile : dict
+            Geotiff profile (attributes)
+        values : ndarray
+            Geotiff data
+        """
+        out_dir = os.path.dirname(geotiff)
+        if not os.path.exists(out_dir):
+            logger.debug("Creating {}".format(out_dir))
+            os.makedirs(out_dir)
+
+        if values.shape[0] != 1:
+            values = np.expand_dims(values, 0)
+
+        dtype = values.dtype.name
+        profile['dtype'] = dtype
+        if np.issubdtype(dtype, np.integer):
+            dtype_max = np.iinfo(dtype).max
+        else:
+            dtype_max = np.finfo(dtype).max
+
+        profile['nodata'] = dtype_max
+
+        with rasterio.open(geotiff, 'w', **profile) as f:
+            f.write(values)
+            logger.debug('\t- {} created'.format(geotiff))
+
+    @classmethod
+    def _extract_layer(cls, excl_h5, layer, geotiff=None, hsds=False):
+        """
+        Extract given layer from exclusions .h5 file and write to geotiff .tif
+
+        Parameters
+        ----------
+        excl_h5 : str
+            Path to .h5 file containing exclusion layers
+        layer : str
+            Layer to extract
+        geotiff : str
+            Path to geotiff file
+        hsds : bool
+            Boolean flag to use h5pyd to handle .h5 'files' hosted on AWS
+            behind HSDS
+
+        Returns
+        -------
+        profile : dict
+            Geotiff profile (attributes)
+        values : ndarray
+            Geotiff data
+        """
+        logger.debug('\t - Extracting {} from {}'
+                     .format(layer, os.path.basename(excl_h5)))
+        with ExclusionLayers(excl_h5, hsds=hsds) as f:
+            profile = f.get_layer_profile(layer)
+            values = f.get_layer_values(layer)
+
+        if geotiff is not None:
+            logger.debug('\t- Writing {} to {}'.format(layer, geotiff))
+            cls._write_geotiff(geotiff, profile, values)
+
+        return profile, values
+
+    def geotiff_to_layer(self, layer, geotiff, check_tiff=True,
+                         transform_atol=0.01, coord_atol=0.001,
+                         description=None, scale_factor=None, dtype='int16'):
+        """
+        Transfer geotiff exclusions to h5 confirming they match existing layers
+
+        Parameters
+        ----------
+        layer : str
+            Layer to extract
+        geotiff : str
+            Path to geotiff file
+        check_tiff : bool, optional
+            Flag to check tiff profile and coordinates against exclusion .h5
+            profile and coordinates, by default True
+        transform_atol : float, optional
+            Absolute tolerance parameter when comparing geotiff transform data,
+            by default 0.01
+        coord_atol : float, optional
+            Absolute tolerance parameter when comparing new un-projected
+            geotiff coordinates against previous coordinates, by default 0.001
+        description : str, optional
+            Description of exclusion layer, by default None
+        scale_factor : int | float, optional
+            Scale factor to use to scale geotiff data when added to the .h5
+            file, by default None
+        dtype : str, optional
+            Dtype to save geotiff data as in the .h5 file. Only used when
+            'scale_factor' is not None, by default 'int16'
+        """
+        if not os.path.exists(self._excl_h5):
+            self._init_h5(self._excl_h5, geotiff, chunks=self._chunks)
+
+        if layer in self.layers:
+            msg = ("{} is already present in {}"
+                   .format(layer, self._excl_h5))
+            if self._replace:
+                msg += " and will be replaced"
+                logger.warning(msg)
+                warn(msg)
+            else:
+                msg += ", to 'replace' set to True"
+                logger.error(msg)
+                raise KeyError(msg)
+
+        self._geotiff_to_h5(self._excl_h5, layer, geotiff,
+                            chunks=self._chunks,
+                            check_tiff=check_tiff,
+                            transform_atol=transform_atol,
+                            coord_atol=coord_atol,
+                            description=description,
+                            scale_factor=scale_factor,
+                            dtype=dtype)
+
+    def layer_to_geotiff(self, layer, geotiff):
+        """
+        Extract desired layer from .h5 file and write to geotiff .tif
+
+        Parameters
+        ----------
+        layer : str
+            Layer to extract
+        geotiff : str
+            Path to geotiff file
+        """
+        self._extract_layer(self._excl_h5, layer, geotiff=geotiff,
+                            hsds=self._hsds)
+
+    @classmethod
+    def layers_to_h5(cls, excl_h5, layers, chunks=(128, 128),
+                     replace=True, check_tiff=True,
+                     transform_atol=0.01, coord_atol=0.001,
+                     descriptions=None, scale_factors=None):
+        """
+        Create exclusions .h5 file, or load layers into existing exclusion .h5
+        file from provided geotiffs
+
+        Parameters
+        ----------
+        excl_h5 : str
+            Path to .h5 file containing or to contain exclusion layers
+        layers : list | dict
+            List of geotiffs to load
+            or dictionary mapping goetiffs to the layers to load
+        chunks : tuple, optional
+            Chunk size of exclusions in Geotiff, by default (128, 128)
+        replace : bool, optional
+            Flag to replace existing layers if needed, by default True
+        check_tiff : bool, optional
+            Flag to check tiff profile and coordinates against exclusion .h5
+            profile and coordinates, by default True
+        transform_atol : float, optional
+            Absolute tolerance parameter when comparing geotiff transform data,
+            by default 0.01
+        coord_atol : float, optional
+            Absolute tolerance parameter when comparing new un-projected
+            geotiff coordinates against previous coordinates, by default 0.001
+        description : dict, optional
+            Description of exclusion layers, by default None
+        scale_factor : dict, optional
+            Scale factors and dtypes to use when scaling given layers,
+            by default None
+        """
+        if isinstance(layers, list):
+            layers = {os.path.basename(lyr).split('.')[0]: lyr
+                      for lyr in layers}
+
+        if descriptions is None:
+            descriptions = {}
+
+        if scale_factors is None:
+            scale_factors = {}
+
+        excls = cls(excl_h5, chunks=chunks, replace=replace)
+        logger.info('Creating {}'.format(excl_h5))
+        for layer, geotiff in layers.items():
+            logger.info('- Transfering {}'.format(layer))
+            description = descriptions.get(layer, None)
+            scale = scale_factors.get(layer, None)
+            if scale is not None:
+                scale_factor = scale['scale_factor']
+                dtype = scale['dtype']
+            else:
+                scale_factor = None
+                dtype = None
+
+            excls.geotiff_to_layer(layer, geotiff, check_tiff=check_tiff,
+                                   transform_atol=transform_atol,
+                                   coord_atol=coord_atol,
+                                   description=description,
+                                   scale_factor=scale_factor,
+                                   dtype=dtype)
+
+    @classmethod
+    def extract_layers(cls, excl_h5, layers, chunks=(128, 128),
+                       hsds=False):
+        """
+        Extract given layers from exclusions .h5 file and save to disk
+        as GeoTiffs
+
+        Parameters
+        ----------
+        excl_h5 : str
+            Path to .h5 file containing or to contain exclusion layers
+        layers : dict
+            Dictionary mapping layers to geotiffs to create
+        chunks : tuple
+            Chunk size of exclusions in .h5 and Geotiffs
+        hsds : bool
+            Boolean flag to use h5pyd to handle .h5 'files' hosted on AWS
+            behind HSDS
+        """
+        excls = cls(excl_h5, chunks=chunks, hsds=hsds)
+        logger.info('Extracting layers from {}'.format(excl_h5))
+        for layer, geotiff in layers.items():
+            logger.info('- Extracting {}'.format(geotiff))
+            excls.layer_to_geotiff(layer, geotiff)
+
+    @classmethod
+    def extract_all_layers(cls, excl_h5, out_dir, chunks=(128, 128),
+                           hsds=False):
+        """
+        Extract all layers from exclusions .h5 file and save to disk
+        as GeoTiffs
+
+        Parameters
+        ----------
+        excl_h5 : str
+            Path to .h5 file containing or to contain exclusion layers
+        out_dir : str
+            Path to output directory into which layers should be saved as
+            GeoTiffs
+        chunks : tuple
+            Chunk size of exclusions in .h5 and Geotiffs
+        hsds : bool
+            Boolean flag to use h5pyd to handle .h5 'files' hosted on AWS
+            behind HSDS
+        """
+        if not os.path.exists(out_dir):
+            os.makedirs(out_dir)
+
+        excls = cls(excl_h5, chunks=chunks, hsds=hsds)
+        logger.info('Extracting layers from {}'.format(excl_h5))
+        for layer in excls.layers:
+            geotiff = os.path.join(out_dir, "{}.tif".format(layer))
+            logger.info('- Extracting {}'.format(geotiff))
+            excls.layer_to_geotiff(layer, geotiff)
+
+
+class Regulations:
+    """Exclusion Regulations. """
+
+    REQUIRED_COLUMNS = ["Feature Type", "Value Type", "Value", "FIPS"]
+
+    def __init__(self, base_regulation_value, regulations_fpath=None,
+                 multiplier=None):
+        """
+        Parameters
+        ----------
+        base_regulation_value : float | int
+            Base regulation value. This value will be used to calculate
+            the exclusion regulation value (e.g. setback distance,
+            flicker hours, etc.) if a multiplier is provided either via
+            the `regulations_fpath`csv or the `multiplier` input. In
+            these cases, the exclusion regulation value will be
+            set to `base_regulation_value * multiplier`.
+        regulations_fpath : str | None, optional
+            Path to regulations .csv or .gpkg file. At a minimum, this
+            file must contain the following columns: `Feature Type`
+            which labels the type of exclusion that each row represents,
+            `Value Type`, which specifies wether the value is a
+            multiplier or static height, `Value`, which specifies the
+            numeric value of the exclusion or multiplier, and `FIPS`,
+            which specifies a unique 5-digit code for each county (this
+            can be an integer - no leading zeros required). Valid
+            options for the `Value Type` are:
+                - "Structure Height Multiplier"
+                - "Meters"
+            If this input is `None`, a generic regulation value of
+            `base_regulation_value * multiplier` is used. By default
+            `None`.
+        multiplier : int | float | str | None, optional
+            A regulation value multiplier to use if regulations are not
+            supplied. This multiplier will be applied to the
+            ``base_regulation_value`` to calculate the exclusion
+            regulation value. If supplied along with
+            ``regulations_fpath``, this input will be used to calculate
+            exclusions for all counties not listed in the regulations
+            file. By default `None`.
+        """
+        self._base_regulation_value = base_regulation_value
+        self._regulations = None
+        self._multi = multiplier
+        self._preflight_check(regulations_fpath)
+
+    def _preflight_check(self, regulations_fpath):
+        """Apply preflight checks to the regulations path and multiplier.
+
+        Run preflight checks on exclusion inputs:
+        1) Ensure either a regulations .csv or
+           an exclusion value multiplier (or both) is provided
+        2) Ensure regulations has county FIPS, map regulations to county
+           geometries from exclusions .h5 file
+
+        Parameters
+        ----------
+        regulations_fpath : str | None
+            Path to regulations .csv file, if `None`, create global
+            exclusions.
+        """
+        if regulations_fpath:
+            try:
+                self.regulations = parse_table(regulations_fpath)
+            except ValueError:
+                self.regulations = gpd.read_file(regulations_fpath)
+            logger.debug('Computing exclusions using regulations provided '
+                         'in: {}'.format(regulations_fpath))
+
+        if self._multi:
+            logger.debug('Computing exclusions using base regulation value '
+                         'multiplier of {}'.format(self._multi))
+
+        if not regulations_fpath and not self._multi:
+            msg = ('Computing exclusions requires a regulations '
+                   '.csv file and/or a generic multiplier!')
+            logger.error(msg)
+            raise RuntimeError(msg)
+
+    @property
+    def regulations(self):
+        """Regulations table.
+
+        Returns
+        -------
+        geopandas.GeoDataFrame | None
+        """
+        return self._regulations
+
+    @regulations.setter
+    def regulations(self, regulations):
+        if regulations is None:
+            msg = "Cannot set regulations to `None`"
+            logger.error(msg)
+            raise ValueError(msg)
+        self._regulations = regulations
+        self._validate_regulations()
+
+    def _validate_regulations(self):
+        """Perform several validations on regulations"""
+
+        self._convert_cols_to_title()
+        self._check_for_req_missing_cols()
+        self._remove_nans_from_req_cols()
+        self._casefold_feature_types()
+
+    def _convert_cols_to_title(self):
+        """Convert column names in regulations DataFrame to str.title(). """
+        new_col_names = {col: col.lower().title()
+                         for col in self._regulations.columns
+                         if col.lower() not in {"geometry", "fips"}}
+        self._regulations = self._regulations.rename(new_col_names, axis=1)
+
+    def _check_for_req_missing_cols(self):
+        """Check for missing (required) columns in regulations DataFrame. """
+        missing = [col for col in self.REQUIRED_COLUMNS
+                   if col not in self._regulations]
+        if any(missing):
+            msg = ('Regulations are missing the following required columns: {}'
+                   .format(missing))
+            logger.error(msg)
+            raise RuntimeError(msg)
+
+    def _remove_nans_from_req_cols(self):
+        """Remove rows with NaN values from required columns. """
+        for col in self.REQUIRED_COLUMNS:
+            na_rows = self._regulations[col].isna()
+            self._regulations = self._regulations[~na_rows]
+
+    def _casefold_feature_types(self):
+        """Casefold "Feature Type" values. """
+        feature_types = self._regulations['Feature Type'].str.strip()
+        feature_types = feature_types.str.casefold()
+        self._regulations['Feature Type'] = feature_types
+
+    @property
+    def base_regulation_value(self):
+        """The base regulation value.
+
+        Returns
+        -------
+        int | float
+        """
+        return self._base_regulation_value
+
+    @property
+    def generic(self):
+        """Default regulation value.
+
+        This value is used for global regulations.
+
+        Returns
+        -------
+        float | None
+        """
+        if self.multiplier is None:
+            regulation_value = None
+        else:
+            regulation_value = self.base_regulation_value * self.multiplier
+
+        return regulation_value
+
+    @property
+    def multiplier(self):
+        """Generic exclusion value multiplier.
+
+        Returns
+        -------
+        int | float
+        """
+        return self._multi
+
+    @property
+    def locals_exist(self):
+        """Flag indicating wether local regulations exist.
+
+        Returns
+        -------
+        bool
+        """
+        return (self.regulations is not None and not self.regulations.empty)
+
+    @property
+    def generic_exists(self):
+        """Flag indicating wether generic regulations exist.
+
+        Returns
+        -------
+        bool
+        """
+        return self.generic is not None
+
+    def __iter__(self):
+        if self._regulations is None:
+            return
+        for ind, county_regulations in self.regulations.iterrows():
+            reg = self._county_regulation_value(county_regulations)
+            if reg is None:
+                continue
+            yield reg, self.regulations.iloc[[ind]].copy()
+
+    def _county_regulation_value(self, county_regulations):
+        """Retrieve county exclusion regulation. """
+        exclusion_type = county_regulations["Value Type"].strip()
+        reg = float(county_regulations["Value"])
+        if exclusion_type.lower() == "structure height multiplier":
+            reg *= self.base_regulation_value
+        elif exclusion_type.lower() != "meters":
+            msg = ("Cannot create exclusions for {}, expecting "
+                   '"Meters", but got {!r}'
+                   .format(county_regulations["County"], exclusion_type))
+            logger.warning(msg)
+            warn(msg)
+            return
+        return reg
