+++ tests/test_setbacks.py
@@ -252,7 +252,7 @@ def test_setbacks_no_computation(setbacks_class):
         regs = Regulations(10, regulations_fpath=regs_fpath)
         setbacks = setbacks_class(EXCL_H5, regs)
         with pytest.raises(ValueError):
-            setbacks.compute_setbacks("RhodeIsland.file")
+            setbacks.compute_exclusions("RhodeIsland.file")
 
 
 @pytest.mark.parametrize(
@@ -266,7 +266,7 @@ def test_setbacks_no_generic_value(setbacks_class, feature_file):
     """Test setbacks computation for invalid input. """
     regs = Regulations(0, regulations_fpath=None, multiplier=1)
     setbacks = setbacks_class(EXCL_H5, regs)
-    out = setbacks.compute_setbacks(feature_file)
+    out = setbacks.compute_exclusions(feature_file)
     assert np.isclose(out, 0).all()
 
 
@@ -282,7 +282,7 @@ def test_generic_structure(generic_wind_regulations):
     setbacks = StructureSetbacks(EXCL_H5, generic_wind_regulations)
     structure_path = os.path.join(TESTDATADIR, 'setbacks',
                                   'RhodeIsland.geojson')
-    test = setbacks.compute_setbacks(structure_path)
+    test = setbacks.compute_exclusions(structure_path)
 
     assert np.allclose(baseline, test)
 
@@ -293,7 +293,7 @@ def test_generic_structure_gpkg(generic_wind_regulations):
     """
     setbacks = StructureSetbacks(EXCL_H5, generic_wind_regulations)
     structure_path = os.path.join(TESTDATADIR, 'setbacks', 'RhodeIsland.gpkg')
-    test = setbacks.compute_setbacks(structure_path)
+    test = setbacks.compute_exclusions(structure_path)
 
     assert test.sum() == 6830
 
@@ -306,14 +306,18 @@ def test_local_structures(max_workers, county_wind_regulations_gpkg):
     baseline = os.path.join(TESTDATADIR, 'setbacks',
                             'existing_structures.tif')
     with Geotiff(baseline) as tif:
-        baseline = tif.values
+        baseline = tif.values[0]
 
     setbacks = StructureSetbacks(EXCL_H5, county_wind_regulations_gpkg)
     structure_path = os.path.join(TESTDATADIR, 'setbacks',
                                   'RhodeIsland.geojson')
-    test = setbacks.compute_setbacks(structure_path, max_workers=max_workers)
+    test = setbacks.compute_exclusions(structure_path, max_workers=max_workers)
 
-    assert np.allclose(baseline, test)
+    # baseline was generated when code did not clip to county bounds,
+    # so test should be a subset of baseline
+    assert baseline.sum() > test.sum()
+    assert (baseline[test > 0] == 1).all()
+    assert (test[baseline == 0] == 0).all()
 
 
 @pytest.mark.parametrize('rail_path',
@@ -330,7 +334,7 @@ def test_generic_railroads(rail_path, generic_wind_regulations):
         baseline = tif.values
 
     setbacks = RailSetbacks(EXCL_H5, generic_wind_regulations)
-    test = setbacks.compute_setbacks(rail_path)
+    test = setbacks.compute_exclusions(rail_path)
 
     assert np.allclose(baseline, test)
 
@@ -342,14 +346,18 @@ def test_local_railroads(max_workers, county_wind_regulations_gpkg):
     """
     baseline = os.path.join(TESTDATADIR, 'setbacks', 'existing_rails.tif')
     with Geotiff(baseline) as tif:
-        baseline = tif.values
+        baseline = tif.values[0]
 
     setbacks = RailSetbacks(EXCL_H5, county_wind_regulations_gpkg)
     rail_path = os.path.join(TESTDATADIR, 'setbacks', 'RI_Railroads',
                              'RI_Railroads.shp')
-    test = setbacks.compute_setbacks(rail_path, max_workers=max_workers)
+    test = setbacks.compute_exclusions(rail_path, max_workers=max_workers)
 
-    assert np.allclose(baseline, test)
+    # baseline was generated when code did not clip to county bounds,
+    # so test should be a subset of baseline
+    assert baseline.sum() > test.sum()
+    assert (baseline[test > 0] == 1).all()
+    assert (test[baseline == 0] == 0).all()
 
 
 def test_generic_parcels():
@@ -359,11 +367,11 @@ def test_generic_parcels():
                                'Rhode_Island.gpkg')
     regulations_x1 = Regulations(BASE_SETBACK_DIST, multiplier=1)
     setbacks_x1 = ParcelSetbacks(EXCL_H5, regulations_x1)
-    test_x1 = setbacks_x1.compute_setbacks(parcel_path)
+    test_x1 = setbacks_x1.compute_exclusions(parcel_path)
 
     regulations_x100 = Regulations(BASE_SETBACK_DIST, multiplier=100)
     setbacks_x100 = ParcelSetbacks(EXCL_H5, regulations_x100)
-    test_x100 = setbacks_x100.compute_setbacks(parcel_path)
+    test_x100 = setbacks_x100.compute_exclusions(parcel_path)
 
     # when the setbacks are so large that they span the entire parcels,
     # a total of 438 regions should be excluded for this particular
@@ -386,11 +394,11 @@ def test_generic_parcels_with_invalid_shape_input():
     setbacks = ParcelSetbacks(EXCL_H5, regulations)
 
     # Ensure data we are using contains invalid shapes
-    parcels = setbacks._parse_features(parcel_path)
+    parcels = setbacks.parse_features(parcel_path)
     assert not parcels.geometry.is_valid.any()
 
     # This code would throw an error if invalid shape not handled properly
-    test = setbacks.compute_setbacks(parcel_path)
+    test = setbacks.compute_exclusions(parcel_path)
 
     # add a test for expected output
     assert not test.any()
@@ -418,7 +426,8 @@ def test_local_parcels_solar(max_workers, regulations_fpath):
 
         parcel_path = os.path.join(TESTDATADIR, 'setbacks', 'RI_Parcels',
                                    'Rhode_Island.gpkg')
-        test = setbacks.compute_setbacks(parcel_path, max_workers=max_workers)
+        test = setbacks.compute_exclusions(parcel_path,
+                                           max_workers=max_workers)
 
     assert test.sum() == 3
 
@@ -462,7 +471,8 @@ def test_local_parcels_wind(max_workers, regulations_fpath):
 
         parcel_path = os.path.join(TESTDATADIR, 'setbacks', 'RI_Parcels',
                                    'Rhode_Island.gpkg')
-        test = setbacks.compute_setbacks(parcel_path, max_workers=max_workers)
+        test = setbacks.compute_exclusions(parcel_path,
+                                           max_workers=max_workers)
 
     assert test.sum() == 3
 
@@ -494,11 +504,11 @@ def test_generic_water_setbacks(water_path):
 
     regulations_x1 = Regulations(BASE_SETBACK_DIST, multiplier=1)
     setbacks_x1 = WaterSetbacks(EXCL_H5, regulations_x1)
-    test_x1 = setbacks_x1.compute_setbacks(water_path)
+    test_x1 = setbacks_x1.compute_exclusions(water_path)
 
     regulations_x100 = Regulations(BASE_SETBACK_DIST, multiplier=100)
     setbacks_x100 = WaterSetbacks(EXCL_H5, regulations_x100)
-    test_x100 = setbacks_x100.compute_setbacks(water_path)
+    test_x100 = setbacks_x100.compute_exclusions(water_path)
 
     # A total of 88,994 regions should be excluded for this particular
     # Rhode Island subset
@@ -531,7 +541,7 @@ def test_local_water_solar(max_workers, regulations_fpath):
 
         water_path = os.path.join(TESTDATADIR, 'setbacks', 'RI_Water',
                                   'Rhode_Island.shp')
-        test = setbacks.compute_setbacks(water_path, max_workers=max_workers)
+        test = setbacks.compute_exclusions(water_path, max_workers=max_workers)
 
     assert test.sum() == 83
 
@@ -571,7 +581,7 @@ def test_local_water_wind(max_workers, regulations_fpath):
 
         water_path = os.path.join(TESTDATADIR, 'setbacks', 'RI_Water',
                                   'Rhode_Island.shp')
-        test = setbacks.compute_setbacks(water_path, max_workers=max_workers)
+        test = setbacks.compute_exclusions(water_path, max_workers=max_workers)
 
     assert test.sum() == 83
 
@@ -650,8 +660,8 @@ def test_partial_exclusions():
     setbacks_hr = ParcelSetbacks(EXCL_H5, regulations,
                                  weights_calculation_upscale_factor=mult)
 
-    exclusion_mask = setbacks.compute_setbacks(parcel_path)
-    inclusion_weights = setbacks_hr.compute_setbacks(parcel_path)
+    exclusion_mask = setbacks.compute_exclusions(parcel_path)
+    inclusion_weights = setbacks_hr.compute_exclusions(parcel_path)
 
     assert exclusion_mask.shape == inclusion_weights.shape
     assert (inclusion_weights < 1).any()
@@ -673,8 +683,8 @@ def test_partial_exclusions_upscale_factor_less_than_1(mult):
     setbacks_hr = ParcelSetbacks(EXCL_H5, regulations,
                                  weights_calculation_upscale_factor=mult)
 
-    exclusion_mask = setbacks.compute_setbacks(parcel_path)
-    inclusion_weights = setbacks_hr.compute_setbacks(parcel_path)
+    exclusion_mask = setbacks.compute_exclusions(parcel_path)
+    inclusion_weights = setbacks_hr.compute_exclusions(parcel_path)
 
     assert np.isclose(exclusion_mask, inclusion_weights).all()
 
@@ -687,7 +697,7 @@ def test_partial_exclusions_upscale_factor_less_than_1(mult):
       REGS_GPKG, 332_887, 142, [HUB_HEIGHT, ROTOR_DIAMETER]),
      (RailSetbacks, WindRegulations,
       os.path.join(TESTDATADIR, 'setbacks', 'Rhode_Island_Railroads.gpkg'),
-      REGS_GPKG, 754_082, 9_402, [HUB_HEIGHT, ROTOR_DIAMETER]),
+      REGS_GPKG, 754_082, 9_276, [HUB_HEIGHT, ROTOR_DIAMETER]),
      (ParcelSetbacks, WindRegulations,
       os.path.join(TESTDATADIR, 'setbacks', 'RI_Parcels', 'Rhode_Island.gpkg'),
       PARCEL_REGS_FPATH_VALUE, 474, 3, [HUB_HEIGHT, ROTOR_DIAMETER]),
@@ -699,7 +709,7 @@ def test_partial_exclusions_upscale_factor_less_than_1(mult):
       REGS_FPATH, 260_963, 112, [BASE_SETBACK_DIST + 199]),
      (RailSetbacks, Regulations,
       os.path.join(TESTDATADIR, 'setbacks', 'Rhode_Island_Railroads.gpkg'),
-      REGS_FPATH, 5_355, 194, [BASE_SETBACK_DIST]),
+      REGS_FPATH, 5_355, 163, [BASE_SETBACK_DIST]),
      (ParcelSetbacks, Regulations,
       os.path.join(TESTDATADIR, 'setbacks', 'RI_Parcels', 'Rhode_Island.gpkg'),
       PARCEL_REGS_FPATH_VALUE, 438, 3, [BASE_SETBACK_DIST]),
@@ -716,7 +726,7 @@ def test_merged_setbacks(setbacks_class, regulations_class, features_path,
                                     multiplier=100)
     generic_setbacks = setbacks_class(EXCL_H5, regulations,
                                       weights_calculation_upscale_factor=sf)
-    generic_layer = generic_setbacks.compute_setbacks(features_path,
+    generic_layer = generic_setbacks.compute_exclusions(features_path,
                                                       max_workers=1)
 
     with tempfile.TemporaryDirectory() as td:
@@ -730,18 +740,18 @@ def test_merged_setbacks(setbacks_class, regulations_class, features_path,
         local_setbacks = setbacks_class(EXCL_H5, regulations,
                                         weights_calculation_upscale_factor=sf)
 
-        local_layer = local_setbacks.compute_setbacks(features_path,
-                                                      max_workers=1)
+        local_layer = local_setbacks.compute_exclusions(features_path,
+                                                        max_workers=1)
 
         regulations = regulations_class(*setback_distance,
                                         regulations_fpath=regs_fpath,
                                         multiplier=100)
         merged_setbacks = setbacks_class(EXCL_H5, regulations,
                                          weights_calculation_upscale_factor=sf)
-        merged_layer = merged_setbacks.compute_setbacks(features_path,
-                                                        max_workers=1)
+        merged_layer = merged_setbacks.compute_exclusions(features_path,
+                                                          max_workers=1)
 
-        local_setbacks._pre_process_regulations(features_path)
+        local_setbacks.pre_process_regulations(features_path)
         feats = local_setbacks.regulations_table
 
     # make sure the comparison layers match what we expect
@@ -811,8 +821,8 @@ def test_merged_setbacks_missing_local(setbacks_class, regulations_class,
     regulations = regulations_class(*setback_distance, regulations_fpath=None,
                                     multiplier=100)
     generic_setbacks = setbacks_class(EXCL_H5, regulations)
-    generic_layer = generic_setbacks.compute_setbacks(features_path,
-                                                      max_workers=1)
+    generic_layer = generic_setbacks.compute_exclusions(features_path,
+                                                        max_workers=1)
 
     with tempfile.TemporaryDirectory() as td:
         regs = pd.read_csv(regulations_fpath).iloc[0:0]
@@ -825,14 +835,14 @@ def test_merged_setbacks_missing_local(setbacks_class, regulations_class,
                                         multiplier=None)
         local_setbacks = setbacks_class(EXCL_H5, regulations)
         with pytest.raises(ValueError):
-            local_setbacks.compute_setbacks(features_path, max_workers=1)
+            local_setbacks.compute_exclusions(features_path, max_workers=1)
 
         regulations = regulations_class(*setback_distance,
                                         regulations_fpath=regs_fpath,
                                         multiplier=100)
         merged_setbacks = setbacks_class(EXCL_H5, regulations)
-        merged_layer = merged_setbacks.compute_setbacks(features_path,
-                                                        max_workers=1)
+        merged_layer = merged_setbacks.compute_exclusions(features_path,
+                                                          max_workers=1)
 
     # make sure the comparison layers match what we expect
     assert generic_layer.sum() == generic_sum
